+++ android/src/main/java/com/mixpanel/reactnative/MixpanelReactNativeModule.java
@@ -1,6 +1,7 @@
 package com.mixpanel.reactnative;
 
 import com.mixpanel.android.mpmetrics.MixpanelAPI;
+import com.mixpanel.android.mpmetrics.MixpanelOptions;
 
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
@@ -35,11 +36,22 @@ public class MixpanelReactNativeModule extends ReactContextBaseJavaModule {
     public void initialize(String token, boolean trackAutomaticEvents, boolean optOutTrackingDefault, ReadableMap metadata, String serverURL, boolean useGzipCompression, Promise promise) throws JSONException {
         JSONObject mixpanelProperties = ReactNativeHelper.reactToJSON(metadata);
         AutomaticProperties.setAutomaticProperties(mixpanelProperties);
-        MixpanelAPI instance = MixpanelAPI.getInstance(this.mReactContext, token, optOutTrackingDefault, mixpanelProperties, null, trackAutomaticEvents);
-        instance.setServerURL(serverURL);
-        if (useGzipCompression) {
-            instance.setShouldGzipRequestPayload(true);
+
+        MixpanelOptions.Builder optionsBuilder = new MixpanelOptions.Builder()
+                .optOutTrackingDefault(optOutTrackingDefault)
+                .superProperties(mixpanelProperties);
+
+        if (serverURL != null && !serverURL.isEmpty()) {
+            optionsBuilder.serverURL(serverURL);
         }
+
+        MixpanelAPI instance = MixpanelAPI.getInstance(this.mReactContext, token, trackAutomaticEvents, optionsBuilder.build());
+
+        if (serverURL != null && !serverURL.isEmpty()) {
+            instance.setServerURL(serverURL);
+        }
+
+        instance.setShouldGzipRequestPayload(useGzipCompression);
         promise.resolve(null);
     }